+++ lib/ogr/driver_mixins/capability_methods.rb
@@ -6,12 +6,17 @@ module OGR
     module CapabilityMethods
       # @return [Boolean] +true+ if this driver supports creating data sources.
       def can_create_data_source?
-        test_capability('CreateDataSource')
+        if GDAL.major_version >= 2
+          # This is weird, but it doesn't work with the normal const.
+          test_capability('DCAP_CREATE')
+        else
+          test_capability(FFI::OGR::Core::ODR_C_CREATE_DATA_SOURCE)
+        end
       end
 
       # @return [Boolean] +true+ if this driver supports deleting data sources.
       def can_delete_data_source?
-        test_capability('DataDataSource')
+        test_capability(FFI::OGR::Core::ODR_C_DELETE_DATA_SOURCE)
       end
     end
   end